+++ packages/nextjs-mf/src/plugins/NextFederationPlugin/index.ts
@@ -84,7 +84,15 @@ export class NextFederationPlugin {
       getModuleFederationPluginConstructor(isServer, compiler);
 
     const defaultShared = retrieveDefaultShared(isServer);
-
+    if (compiler.options.output.chunkFilename) {
+      // @ts-ignore
+      compiler.options.output.chunkFilename =
+        // @ts-ignore
+        compiler.options.output.chunkFilename.replace(
+          'contenthash',
+          'fullhash'
+        );
+    }
     if (isServer) {
       // Refactored server condition
       configureServerCompilerOptions(compiler);
@@ -99,14 +107,15 @@ export class NextFederationPlugin {
       applyClientPlugins(compiler, this._options, this._extraOptions);
     }
     //@ts-ignore
-    compiler.options.module.parser.javascript.importMeta = false;
+    //compiler.options.module.parser.javascript.importMeta = false;
 
     // @ts-ignore
     const hostFederationPluginOptions: ModuleFederationPluginOptions = {
       ...this._options,
       runtime: false,
       exposes: {
-        // __hoist: require.resolve('../../delegate-hoist-container'),
+        //something must be exposed in order to generate a remote entry, which is needed to kickstart runtime
+        './noop': require.resolve('../../federation-noop.js'),
         ...(this._extraOptions.exposePages
           ? exposeNextjsPages(compiler.options.context as string)
           : {}),